Move service and remove guava dependency
